Move TNumberTheory to AdvancedMath instead of its own namespace
