+++ TMath.ConsoleApp/Program.cs
@@ -1,14 +1,15 @@
 using System.Numerics;
 using TMath.Numerics;
+using TMath.Numerics.AdvancedMath;
 using TMath.Numerics.AdvancedMath.LinearAlgebra;
-using TMath.Numerics.AdvancedMath.NumberTheory;
 namespace TMath.ConsoleApp
 {
     internal class Program
     {
         static void Main(string[] args)
         {
-            Console.WriteLine(string.Join(", ", TNumberTheory.GeneratePrimesUpTo(542)));
+            int[] arr = null;
+            Console.WriteLine(TNumberTheory.CollatzConjecture(-1).Select(x => $"{x}, ").Aggregate((a, b) => a + b));
         }
     }
 }
